dom/media/mediasource/test/test_ResumeAfterClearing_mp4.html

<!DOCTYPE html>
<html><head>
<meta http-equiv="content-type" content="text/html; charset=windows-1252">
  <title>MSE: Don't get stuck buffering for too long when we have frames to show</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test"><script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(async (ms, v) => {
  await once(ms, "sourceopen");
  ok(true, "Receive a sourceopen event");
  ms.addEventListener("sourceopen", () => ok(false, "No more sourceopen"));
  const sb = ms.addSourceBuffer("video/mp4");
  ok(sb, "Create a SourceBuffer");
  sb.addEventListener("error", e => {
    ok(false, "Got Error: " + e);
    SimpleTest.finish();
  });
  await fetchAndLoad(sb, "bipbop/bipbop", ["init"], ".mp4");
  let p = once(v, "loadeddata");
  await fetchAndLoad(sb, "bipbop/bipbop", range(1, 3), ".m4s");
  await p;
  // clear the entire sourcebuffer.
  sb.remove(0, 5);
  await once(sb, "updateend");
  v.play();
  // We have nothing to play, waiting will be fired.
  await once(v, "waiting");
  p = once(v, "playing");
  await fetchAndLoad(sb, "bipbop/bipbop", range(1, 4), ".m4s");
  await p;
  ms.endOfStream();
  await Promise.all([once(ms, "sourceended"), once(v, "ended")]);
  SimpleTest.finish(SimpleTest);
});
</script>
</pre>
</body>
</html>
